requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2

# Para Railway (produccion)
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
dj-database-url==2.3.0
